+++ internal/provider/resource_tfe_opa_version_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package tfe
+package provider
 
 import (
 	"crypto/hmac"
