+++ internal/provider/resource_tfe_sentinel_version.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package tfe
+package provider
 
 import (
 	"context"
